pyproject.toml
[project]
name = "novi-py"
version = "0.1.0"
description = "Python client for Novi"
authors = [
    { name = "mivik", email = "mivik@qq.com" }
]
dependencies = [
    "uuid>=1.30",
    "grpcio-status>=1.64.0",
    "pydantic>=2.7.1",
    "grpcio>=1.64.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "grpcio-tools>=1.64.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/novi"]
